Added some attributes that flag an IPart as needing required by either a vendor at the point the trade window is opened, or an item in the trade window at the point immediately before vendor actions are collected.
Added a fallback to the base game's implementation of VendorActions if the event returns an empty list.

Switched to "safe" versions of blueprint retrieval to prevent inheritance-error-based crashes on player load.

Added startup code to detect if the mod was added mid-save and to behave appropriately, if so.

Added some warnings to more easily identify which the culprit mod is in the event that there is an inheritance error affecting this mod's processes (without resulting in a crash).